requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
dataclasses-json==0.6.1
defusedxml==0.7.1
exceptiongroup==1.1.3
fastapi==0.104.0
frozenlist==1.4.0
h11==0.14.0
httpie==3.2.2
idna==3.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.325
langchain-experimental==0.0.38
langsmith==0.0.53
markdown-it-py==3.0.0
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.4
openai==0.28.1
packaging==23.2
pandas==2.0.3
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
tabulate==0.9.0
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.23.2
yarl==1.9.2
